B2B Pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
amqp==5.3.1
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
discord.py==2.6.4
exceptiongroup==1.3.1
feedparser==6.0.12
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
SQLAlchemy==2.0.45
streamlit==1.52.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.14
yarl==1.22.0
